requirements.txt
opencv-python
yt-dlp
kafka-python
minio